requirements.txt
pillow
pulsectl